Moved src into myproject
