+++ myproject/src/shortgen/gpt.py
