+++ myproject/src/videoeditor/silencedetect.py
