+++ myproject/src/videoeditor/video.py
@@ -1,8 +1,8 @@
 from moviepy.editor import VideoFileClip, concatenate_videoclips, CompositeVideoClip, ImageClip
 from moviepy.config import change_settings
-from silencedetect import silence_detect
+from videoeditor.silencedetect import silence_detect
 import glob
-from utils import *
+from videoeditor.utils import *
 
 change_settings({"IMAGEMAGICK_BINARY": r"c:\Program Files\ImageMagick-7.1.1-Q16-HDRI\magick.exe"})
